--- src/pages/Login/LogInForm.tsx
@@ -1,5 +1,0 @@
-function LogInForm() {
-  return <h1>로그인 컴포넌트입니다.</h1>;
-}
-
-export default LogInForm;
