--- src/pages/Login/index.tsx
@@ -1,4 +1,0 @@
-function Login() {
-  return <h1>로그인 페이지입니다.</h1>;
-}
-export default Login;
