+++ src/pages/sign/Login/LogInForm.tsx
@@ -1,0 +1,11 @@
+const LoginForm = () => {
+  return (
+    <div>
+      <input type="email" placeholder="이메일" />
+      <input type="password" placeholder="비밀번호" />
+      <button type="submit">로그인</button>
+    </div>
+  );
+};
+
+export default LoginForm;
