+++ src/pages/sign/Login/index.tsx
@@ -1,0 +1,12 @@
+import { default as LoginForm } from './LogInForm';
+
+const Login = () => {
+  return (
+    <section>
+      <h1>로그인</h1>
+      <LoginForm />
+    </section>
+  );
+};
+
+export default Login;
