+++ src/pages/sign/SignUp/SignUpForm.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 const SignUpForm = () => {
   return <div>SignUpForm</div>;
 };
