+++ src/pages/sign/SignUp/index.tsx
@@ -1,11 +1,12 @@
-import React from 'react';
+import { default as SignUpForm } from './SignUpForm';
 
-function SignUp() {
+const SignUp = () => {
   return (
     <div>
       <h1>회원가입 페이지입니다.</h1>
+      <SignUpForm />
     </div>
   );
-}
+};
 
 export default SignUp;
