+++ src/pages/sign/index.ts
@@ -1,0 +1,2 @@
+export { default as Login } from './Login';
+export { default as SignUp } from './SignUp';
